Close #5 refactor: Member의 필드인 Enum 클래스 Member의 내부 클래스로 수정
Grade, Region, Authority Enum 클래스는 Member 도메인 클래스의 하위 필드이기 때문에 Member 내부 클래스로 이동

- MemberMapper.xml, PostMapper.xml ResultMap 수정
- MemberRepository, PostRepository 마이바티스 통합 테스트 완료
